create a different component for Star Rating
